requirements.txt
django
djangorestframework
django-filter
six
Pillow
opencv-python==4.0.0.21
sklearn
scipy
markdown
requests